% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serviceclass.R
\name{ServiceClass}
\alias{ServiceClass}
\title{ServiceClass}
\value{
R6 object
}
\description{
ServiceClass： this is an R6 object for systemd service create and run
}
\examples{
ServiceClass$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Unit}}{character Unit options of service file}

\item{\code{Service}}{character Service options of service file}

\item{\code{Install}}{character Install options of service file}

\item{\code{servicename}}{character name of service}

\item{\code{save_dir}}{character save_dir of created service file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ServiceClass-new}{\code{ServiceClass$new()}}
\item \href{#method-ServiceClass-info}{\code{ServiceClass$info()}}
\item \href{#method-ServiceClass-create}{\code{ServiceClass$create()}}
\item \href{#method-ServiceClass-service_run}{\code{ServiceClass$service_run()}}
\item \href{#method-ServiceClass-service_control}{\code{ServiceClass$service_control()}}
\item \href{#method-ServiceClass-clone}{\code{ServiceClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceClass-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceClass$new(
  Unit = NULL,
  Service = NA,
  Install = c("WantedBy=multi-user.target"),
  servicename = "temple",
  save_dir = getwd()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Unit}}{character Unit options of service file}

\item{\code{Service}}{character Service options of service file}

\item{\code{Install}}{character Install options of service file}

\item{\code{servicename}}{character name of service}

\item{\code{save_dir}}{character save_dir of created service file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceClass-info"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceClass-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceClass$info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceClass-create"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceClass-create}{}}}
\subsection{Method \code{create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceClass$create()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceClass-service_run"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceClass-service_run}{}}}
\subsection{Method \code{service_run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceClass$service_run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceClass-service_control"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceClass-service_control}{}}}
\subsection{Method \code{service_control()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceClass$service_control(
  mod = c("stop", "restart", "enable", "disable", "status")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{the mod of service control}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
